added various methods asking for toppings

+++ src/main/java/com/pluralsight/Toppings/RegularTopping.java
@@ -8,7 +8,7 @@ public class RegularTopping extends Topping {
     }
 
     @Override
-    public double getPrice(double size) {
+    public double getPrice(int size) {
         return 0;
     }
