+++ src/main/java/com/pluralsight/Toppings/Sauce.java
@@ -8,7 +8,7 @@ public class Sauce extends Topping {
     }
 
     @Override
-    public double getPrice(double size) {
+    public double getPrice(int size) {
         return 0;
     }
 }
